test_requirements.txt
django-bootstrap-pagination==1.6.1
django-crispy-forms==1.6.0
sorl-thumbnail==12.3
